fix: Modal component is no longer rendered from App component
